server/requirements.txt
scipy==1.5.4
Flask==1.1.2
tensorflow==2.3.1
Flask_Cors==3.0.9
pandas==1.1.4
numpy==1.18.5
python-dotenv==0.15.0
transformers==3.4.0
tweepy==3.9.0